[hybris] implement adreno quriks:
When --enable-adreno-quirks is specified then treat malloc(4) as
if it was malloc(5) in order to avoid SIGSEGV and shader
compilation errors due to out-of-bounds accesses found within
common adreno EGL/GLES blobs.

These errors are usually found on 64 bit devices and this works
around this issue until a better solution is found.